print the next instruction to be executed, not the one that is currently executing.
This is so that the user can inspect the relevant registers and memory
before and after the execution of the instruction.

I have no idea why a person would want to hit step and then
see the instruction that was just executed.  Therefore, print
out the PC after it has been set to its next value.